pyproject.toml
[tool.poetry]
name = "argos-sct"
version = "0.5.0"
description = "Multi-operator swarm control framework"
authors = ["Genki Miyauchi <4963919-genki_miyauchi@users.noreply.gitlab.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.21.3"
matplotlib = "^3.4.3"
protobuf = "^3.19.0"
PyYAML = "^6.0"
pandas = "^1.4.0"
pyperclip = "^1.8.2"
Flask = "^2.0.3"
websocket-client = "^1.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
